opencode_projects/Daily-News/requirements.txt
feedparser==6.0.10
trafilatura>=1.6.0
#newspaper3k==0.2.8
lxml_html_clean
beautifulsoup4==4.12.2
requests==2.31.0
ollama==0.1.7
PyYAML==6.0.1
